+++ Sources/Mappable.swift
@@ -36,11 +36,30 @@ public protocol Mappable {
      */
     @warn_unused_result
     static func from(JSON: NSDictionary) -> Self?
+
+    /**
+     Convenience method for creating Mappable objects from a NSArray
+
+     - parameter JSON: The JSON to create the objects from
+
+     - returns: An array of the created objects, or nil if creating threw
+     */
+    @warn_unused_result
+    static func from(JSON: NSArray) -> [Self]?
 }
 
 public extension Mappable {
     @warn_unused_result
     public static func from(JSON: NSDictionary) -> Self? {
         return try? self.init(map: Mapper(JSON: JSON))
+    }
+
+    @warn_unused_result
+    public static func from(JSON: NSArray) -> [Self]? {
+        if let array = JSON as? [NSDictionary] {
+            return try? array.map { try self.init(map: Mapper(JSON: $0)) }
+        }
+
+        return nil
     }
 }