+++ Tests/InitializerTests.swift
@@ -12,6 +12,9 @@ extension TestExtension: Mappable {
 }
 
 final class InitializerTests: XCTestCase {
+
+    // MARK: from NSDictionary
+
     func testCreatingInvalidFromJSON() {
         struct Test: Mappable {
             let string: String
@@ -36,7 +39,46 @@ final class InitializerTests: XCTestCase {
         XCTAssertTrue(test?.string == "Hi")
     }
 
-    // Testing http://www.openradar.me/23376350
+    // MARK: from NSArray
+
+    func testCreatingFromArrayOfJSON() {
+        struct Test: Mappable {
+            let string: String
+            init(map: Mapper) throws {
+                try self.string = map.from("string")
+            }
+        }
+
+        let tests = Test.from([["string": "Hi"], ["string": "Bye"]])
+        XCTAssertTrue(tests?.count == 2)
+    }
+
+    func testCreatingFromPartiallyInvalidArrayOfJSON() {
+        struct Test: Mappable {
+            let string: String
+            init(map: Mapper) throws {
+                try self.string = map.from("string")
+            }
+        }
+
+        let tests = Test.from([["string": "Hi"], ["nope": "Bye"]])
+        XCTAssertNil(tests)
+    }
+
+    func testCreatingFromInvalidArray() {
+        struct Test: Mappable {
+            let string: String
+            init(map: Mapper) throws {
+                try self.string = map.from("string")
+            }
+        }
+
+        let tests = Test.from(["hi"])
+        XCTAssertNil(tests)
+    }
+
+    // MARK: Testing http://www.openradar.me/23376350
+
     func testCreatingWithConformanceInExtension() {
         let test = TestExtension.from(["string": "Hi"])
         XCTAssertTrue(test?.string == "Hi")