+++ CHANGELOG.md
@@ -6,7 +6,9 @@
 
 ## Enhancements
 
-- None
+- Add function for creating `Mappable`s from a `NSArray`
+  [Keith Smiley](https://github.com/keith)
+  [#27](https://github.com/lyft/mapper/pull/27)
 
 ## Bug Fixes